Move types to their own module
