+++ 2024/rust/day15/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod part1;
 pub mod part2;
+pub mod types;
